Improvements and fixes to lfs

+++ tests/test_rdd.rs
@@ -1,12 +1,11 @@
 use native_spark::io::*;
+use native_spark::partitioner::HashPartitioner;
+use native_spark::rdd::CoGroupedRdd;
 use native_spark::*;
-
-use std::fs::{create_dir_all, remove_dir_all, File};
+use serde_traitobject::Arc as SerArc;
+use std::fs::{create_dir_all, File};
 use std::io::prelude::*;
 use std::sync::Arc;
-use native_spark::rdd::CoGroupedRdd;
-use native_spark::partitioner::HashPartitioner;
-use serde_traitobject::Arc as SerArc;
 
 #[macro_use]
 extern crate serde_closure;
@@ -16,31 +15,19 @@ static CONTEXT: Lazy<Arc<Context>> = Lazy::new(|| Context::new().unwrap());
 static WORK_DIR: Lazy<std::path::PathBuf> = Lazy::new(std::env::temp_dir);
 const TEST_DIR: &str = "ns_test_dir";
 
+#[allow(unused_must_use)]
 fn set_up(file_name: &str) {
     let temp_dir = WORK_DIR.join(TEST_DIR);
     println!("Creating tests in dir: {}", (&temp_dir).to_str().unwrap());
-    create_dir_all(&temp_dir).unwrap();
+    create_dir_all(&temp_dir);
 
     let fixture =
         b"This is some textual test data.\nCan be converted to strings and there are two lines.";
 
-    let mut f = File::create(temp_dir.join(file_name)).unwrap();
-    f.write_all(fixture).unwrap();
-}
-
-fn tear_down() {
-    // Clean up files
-    let temp_dir = WORK_DIR.join(TEST_DIR);
-    remove_dir_all(temp_dir).unwrap();
-}
-
-fn test_runner<T>(test: T)
-where
-    T: FnOnce() -> () + std::panic::UnwindSafe,
-{
-    let result = std::panic::catch_unwind(|| test());
-    tear_down();
-    assert!(result.is_ok())
+    if !std::path::Path::new(file_name).exists() {
+        let mut f = File::create(temp_dir.join(file_name)).unwrap();
+        f.write_all(fixture).unwrap();
+    }
 }
 
 #[test]
@@ -174,68 +161,72 @@ fn test_first() {
 }
 
 #[test]
-fn test_read_files() -> Result<()> {
-    // Single file test
-    let file_name = "test_file_01";
-    let file_path = WORK_DIR.join(TEST_DIR).join(file_name);
-    set_up(file_name);
-
-    let processor = Fn!(|reader: DistributedLocalReader| {
-        let mut files: Vec<_> = reader.into_iter().collect();
-        assert_eq!(files.len(), 1);
-
+fn test_read_files_bytes() -> Result<()> {
+    let deserializer = Fn!(|file: Vec<u8>| -> Vec<String> {
         // do stuff with the read files ...
-        let parsed: Vec<_> = String::from_utf8(files.pop().unwrap())
+        let parsed: Vec<_> = String::from_utf8(file)
             .unwrap()
             .lines()
             .map(|s| s.to_string())
             .collect();
-
         assert_eq!(parsed.len(), 2);
         assert_eq!(parsed[0], "This is some textual test data.");
-
-        // return parsed stuff
+        // return lines
         parsed
     });
 
-    test_runner(|| {
-        let sc = CONTEXT.clone();
-        let result = sc
-            .read_files(LocalFsReaderConfig::new(file_path), processor)
-            .collect()
-            .unwrap();
-        assert_eq!(result[0].len(), 2);
-    });
+    // Single file test
+    let file_name = "test_file_1";
+    let file_path = WORK_DIR.join(TEST_DIR).join(file_name);
+    set_up(file_name);
+
+    let context = CONTEXT.clone();
+    let result = context
+        .read_files(LocalFsReaderConfig::new(file_path), deserializer)
+        .collect()
+        .unwrap();
+    assert_eq!(result[0].len(), 2);
 
     // Multiple files test
-    let _multi_files: Vec<_> = (0..10)
-        .map(|idx| {
-            let f_name = format!("test_file_{}", idx);
-            let path = WORK_DIR.join(TEST_DIR).join(f_name.as_str());
-            set_up(path.as_path().to_str().unwrap());
-        })
-        .collect::<Vec<_>>();
+    (0..10).for_each(|idx| {
+        let f_name = format!("test_file_{}", idx);
+        let path = WORK_DIR.join(TEST_DIR).join(f_name.as_str());
+        set_up(path.as_path().to_str().unwrap());
+    });
 
-    let processor = Fn!(|reader: DistributedLocalReader| {
-        let files: Vec<_> = reader.into_iter().collect();
+    let sc = CONTEXT.clone();
+    let files = sc.clone().read_files(
+        LocalFsReaderConfig::new(WORK_DIR.join(TEST_DIR)),
+        deserializer,
+    );
+    let result: Vec<_> = files.collect().unwrap().into_iter().flatten().collect();
+    assert_eq!(result.len(), 20);
 
-        // do stuff with the read files ...
-        let parsed: Vec<_> = files
-            .into_iter()
-            .map(|f| String::from_utf8(f).unwrap())
-            .flat_map(|s| s.lines().map(|l| l.to_owned()).collect::<Vec<_>>())
-            .collect();
+    Ok(())
+}
 
-        // return parsed stuff
+#[test]
+fn test_read_files() -> Result<()> {
+    let deserializer = Fn!(|file: std::path::PathBuf| {
+        let mut file = File::open(file).unwrap();
+        let mut content = String::new();
+        file.read_to_string(&mut content).unwrap();
+        let parsed: Vec<_> = content.lines().map(|s| s.to_string()).collect();
+        assert_eq!(parsed.len(), 2);
+        assert_eq!(parsed[0], "This is some textual test data.");
         parsed
     });
 
-    test_runner(|| {
-        let sc = CONTEXT.clone();
-        let files = sc.read_files(LocalFsReaderConfig::new(WORK_DIR.join(TEST_DIR)), processor);
-        let result: Vec<_> = files.collect().unwrap().into_iter().flatten().collect();
-        assert_eq!(result.len(), 20);
-    });
+    let file_name = "test_file_1";
+    let file_path = WORK_DIR.join(TEST_DIR).join(file_name);
+    set_up(file_name);
+    let context = CONTEXT.clone();
+    let result = context
+        .read_files(LocalFsReaderConfig::new(file_path), deserializer)
+        .collect()
+        .unwrap();
+    assert_eq!(result[0].len(), 2);
+
     Ok(())
 }
 
@@ -246,39 +237,36 @@ fn test_distinct() -> Result<()> {
     let rdd = sc
         .clone()
         .parallelize(vec![1, 2, 2, 2, 3, 3, 3, 4, 4, 5], 3);
-    assert!(rdd.distinct().collect()?.len() == 5);
-    assert!(
+    assert_eq!(rdd.distinct().collect()?.len(), 5);
+    assert_eq!(
+        rdd.distinct()
+            .collect()?
+            .into_iter()
+            .collect::<HashSet<_>>(),
         rdd.distinct()
             .collect()?
             .into_iter()
             .collect::<HashSet<_>>()
-            == rdd
-                .distinct()
-                .collect()?
-                .into_iter()
-                .collect::<HashSet<_>>()
     );
-    assert!(
+    assert_eq!(
         rdd.distinct_with_num_partitions(2)
             .collect()?
             .into_iter()
-            .collect::<HashSet<_>>()
-            == rdd
-                .distinct()
-                .collect()?
-                .into_iter()
-                .collect::<HashSet<_>>()
-    );
-    assert!(
-        rdd.distinct_with_num_partitions(10)
+            .collect::<HashSet<_>>(),
+        rdd.distinct()
             .collect()?
             .into_iter()
             .collect::<HashSet<_>>()
-            == rdd
-                .distinct()
-                .collect()?
-                .into_iter()
-                .collect::<HashSet<_>>()
+    );
+    assert_eq!(
+        rdd.distinct_with_num_partitions(10)
+            .collect()?
+            .into_iter()
+            .collect::<HashSet<_>>(),
+        rdd.distinct()
+            .collect()?
+            .into_iter()
+            .collect::<HashSet<_>>()
     );
     Ok(())
 }
@@ -309,6 +297,40 @@ fn test_partition_wise_sampling() -> Result<()> {
         let result = rdd.take_sample(false, 10, None)?;
         assert!(result.len() == 10);
     }
+    Ok(())
+}
+
+#[test]
+fn test_cartesian() -> Result<()> {
+    let sc = CONTEXT.clone();
+    let rdd1 = sc.parallelize((0..2).collect::<Vec<_>>(), 2);
+    let rdd2 = sc.parallelize("αβ".chars().collect::<Vec<_>>(), 2);
+
+    let res = rdd1.cartesian(rdd2).collect()?;
+    itertools::assert_equal(res, vec![(0, 'α'), (0, 'β'), (1, 'α'), (1, 'β')]);
+    Ok(())
+}
+
+#[test]
+fn test_coalesced() -> Result<()> {
+    let sc = CONTEXT.clone();
+
+    // do not shuffle
+    {
+        let rdd = sc.parallelize(vec![1; 101], 101);
+        let res = rdd.coalesce(5, false).glom().collect()?;
+        assert_eq!(res.len(), 5);
+        assert_eq!(res[0].iter().sum::<u8>(), 20);
+        assert_eq!(res[4].iter().sum::<u8>(), 21);
+    }
+
+    // shuffle and increase num partitions
+    {
+        let rdd = sc.parallelize(vec![1; 100], 20);
+        let res = rdd.repartition(100).glom().collect()?;
+        assert_eq!(res.len(), 100);
+    }
+
     Ok(())
 }
 
@@ -350,40 +372,6 @@ fn test_union() -> Result<()> {
 }
 
 #[test]
-fn test_cartesian() -> Result<()> {
-    let sc = CONTEXT.clone();
-    let rdd1 = sc.parallelize((0..2).collect::<Vec<_>>(), 2);
-    let rdd2 = sc.parallelize("αβ".chars().collect::<Vec<_>>(), 2);
-
-    let res = rdd1.cartesian(rdd2).collect()?;
-    itertools::assert_equal(res, vec![(0, 'α'), (0, 'β'), (1, 'α'), (1, 'β')]);
-    Ok(())
-}
-
-#[test]
-fn test_coalesced() -> Result<()> {
-    let sc = CONTEXT.clone();
-
-    // do not shuffle
-    {
-        let rdd = sc.parallelize(vec![1; 101], 101);
-        let res = rdd.coalesce(5, false).glom().collect()?;
-        assert_eq!(res.len(), 5);
-        assert_eq!(res[0].iter().sum::<u8>(), 20);
-        assert_eq!(res[4].iter().sum::<u8>(), 21);
-    }
-
-    // shuffle and increase num partitions
-    {
-        let rdd = sc.parallelize(vec![1; 100], 20);
-        let res = rdd.repartition(100).glom().collect()?;
-        assert_eq!(res.len(), 100);
-    }
-
-    Ok(())
-}
-
-#[test]
 fn test_union_with_unique_partitioner() {
     let sc = CONTEXT.clone();
     let partitioner = HashPartitioner::<i32>::new(2);
@@ -394,10 +382,9 @@ fn test_union_with_unique_partitioner() {
             (3, "C".to_string()),
             (4, "D".to_string()),
         ];
-        let rdd0 = SerArc::new(sc.parallelize(rdd.clone(), 2))
-            as SerArc<dyn Rdd<Item = (i32, String)>>;
-        let rdd1 =
-            SerArc::new(sc.parallelize(rdd, 2)) as SerArc<dyn Rdd<Item = (i32, String)>>;
+        let rdd0 =
+            SerArc::new(sc.parallelize(rdd.clone(), 2)) as SerArc<dyn Rdd<Item = (i32, String)>>;
+        let rdd1 = SerArc::new(sc.parallelize(rdd, 2)) as SerArc<dyn Rdd<Item = (i32, String)>>;
         CoGroupedRdd::<i32>::new(
             vec![rdd0.get_rdd_base().into(), rdd1.get_rdd_base().into()],
             Box::new(partitioner.clone()),
